SparkServiceContainer no longer depending on System.Web.Hosting
- SparkViewEngine no longer depends on System.Web.Hosting
- Removed "ViewFolderType.cs" enum
- Can no longer configure "ViewFolder" using the view folder type, you have to either use the default or configure it one using as if it was a custom one
- New Spark.Web project to contain the bits that depend on System.Web (e.g. VirtualPathProviderViewFolder)
