+++ src/Spark.JsTests/Spark.JsTests.csproj
@@ -13,7 +13,7 @@
 		<ProjectCapability Include="SupportsSystemWeb" />
 	</ItemGroup>
 	<ItemGroup>
-	  <ProjectReference Include="..\Spark\Spark.csproj" />
+	  <ProjectReference Include="..\Spark.Web\Spark.Web.csproj" />
 	</ItemGroup>
 	<ItemGroup>
 	  <Reference Include="System.Web" />
