+++ src/Spark.Web.Mvc/Spark.Web.Mvc.csproj
@@ -34,7 +34,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Spark\Spark.csproj" />
+    <ProjectReference Include="..\Spark.Web\Spark.Web.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\license.txt">
