+++ src/Spark.Web.Tests/App.config
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<configuration>
+  <configSections>
+    <section name="spark" type="Spark.Configuration.SparkSectionHandler, Spark"/>
+  </configSections>
+  <spark>
+    <compilation debug="true" nullBehaviour="Strict">
+      <assemblies>
+        <add assembly="System.Web, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
+      </assemblies>
+    </compilation>
+    <pages pageBaseType="Spark.Tests.Stubs.StubSparkView" automaticEncoding="false">
+      <namespaces>
+        <add namespace="System"/>
+      </namespaces>
+    </pages>
+  </spark>
+</configuration>
