+++ src/Spark.sln
@@ -54,6 +54,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Castle.MonoRail.Pdf", "Cast
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Castle.MonoRail.Pdf.Tests", "Castle.Monorail.Pdf.Tests\Castle.MonoRail.Pdf.Tests.csproj", "{0FEC123E-DFD8-4DB0-9FC7-6C2EF2897FAD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spark.Web", "Spark.Web\Spark.Web.csproj", "{4A68431B-ADD3-4F99-9121-2DF0C5C93BC3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spark.Web.Tests", "Spark.Web.Tests\Spark.Web.Tests.csproj", "{58EE303C-D806-4917-9CA8-C2B10594A1AA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -136,6 +140,14 @@ Global
 		{0FEC123E-DFD8-4DB0-9FC7-6C2EF2897FAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0FEC123E-DFD8-4DB0-9FC7-6C2EF2897FAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0FEC123E-DFD8-4DB0-9FC7-6C2EF2897FAD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A68431B-ADD3-4F99-9121-2DF0C5C93BC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A68431B-ADD3-4F99-9121-2DF0C5C93BC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A68431B-ADD3-4F99-9121-2DF0C5C93BC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A68431B-ADD3-4F99-9121-2DF0C5C93BC3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{58EE303C-D806-4917-9CA8-C2B10594A1AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{58EE303C-D806-4917-9CA8-C2B10594A1AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{58EE303C-D806-4917-9CA8-C2B10594A1AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{58EE303C-D806-4917-9CA8-C2B10594A1AA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -160,6 +172,8 @@ Global
 		{0799EE7A-58D3-467F-8DE6-8F3FFDDEF4A1} = {F4E1F2B2-2BA6-4EFC-8082-ADBC14ECC179}
 		{D5796359-FF68-4444-809D-C7A8DAA92E41} = {CE293579-4889-4D6D-A1A0-BBBE8CBB9343}
 		{0FEC123E-DFD8-4DB0-9FC7-6C2EF2897FAD} = {CE293579-4889-4D6D-A1A0-BBBE8CBB9343}
+		{4A68431B-ADD3-4F99-9121-2DF0C5C93BC3} = {F606A062-1BD5-40E2-8E8A-BCC4F374BFBF}
+		{58EE303C-D806-4917-9CA8-C2B10594A1AA} = {F606A062-1BD5-40E2-8E8A-BCC4F374BFBF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B9FB0A1C-BC73-4CE6-95F0-9D33AE572BED}
